[messages] Moved discrete components into separate files
