+++ src/components/common/Header.tsx
@@ -4,7 +4,7 @@ import AppBar from "@material-ui/core/AppBar";
 import CircularProgress from "@material-ui/core/CircularProgress";
 import Typography from "@material-ui/core/Typography";
 
-import { useStatusReporter } from "~/components/StatusReporter";
+import { useStatusReporter } from "~/components/contexts/StatusReporter";
 
 /**
  * A simple app Header bar, which also shows the Status from StatusReporter
